feat: add Category model and update Lab relationships for category management
