+++ app/domain/lab/model.py
@@ -5,6 +5,7 @@ from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database.connection import Base
 from app.common.audit_mixin import TimestampMixin
+from app.domain.category.model import Category
 
 
 lab_category = Table(
@@ -28,7 +29,7 @@ class Lab(Base, TimestampMixin):
     name: Mapped[str] = mapped_column(String(150), nullable=False)
     focus: Mapped[str | None] = mapped_column(String(255), nullable=True)
     description: Mapped[str | None] = mapped_column(Text, nullable=True)
-    categories: Mapped[list["Category"]] = relationship(
+    categories: Mapped[list[Category]] = relationship(
         "Category",
         secondary=lab_category,
         back_populates="labs",
@@ -39,17 +40,3 @@ class Lab(Base, TimestampMixin):
     @property
     def category_ids(self) -> list[int]:
         return [category.id for category in self.categories]
-
-
-class Category(Base, TimestampMixin):
-    __tablename__ = "categories"
-
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    name: Mapped[str] = mapped_column(String(100), nullable=False, unique=True)
-
-    labs: Mapped[list["Lab"]] = relationship(
-        "Lab",
-        secondary=lab_category,
-        back_populates="categories",
-        lazy="selectin",
-    )
