+++ app/domain/lab/repository.py
@@ -5,7 +5,8 @@ from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.common.base_repository import BaseRepository
-from app.domain.lab.model import Lab, Category
+from app.domain.lab.model import Lab
+from app.domain.category.model import Category
 from app.domain.lab.schema import LabCreateSchema, LabOutSchema
 from app.exceptions.exceptions import DatabaseError, NotFoundError
